+++ src/constant/constant.ts
@@ -1,10 +1,10 @@
 import type { YobtaSyncRule } from '../ruleYobta/index.js'
 import { ruleYobta } from '../ruleYobta/index.js'
 
-export const constMessage = <I>(value: I): string =>
+export const constantMessage = <I>(value: I): string =>
   `Should be identical to "${String(value)}"`
 
-export function constYobta<I>(
+export function constant<I>(
   value: I,
   message?: string,
 ): YobtaSyncRule<unknown, I> {
@@ -12,6 +12,6 @@ export function constYobta<I>(
     if (input === value) {
       return input as I
     }
-    throw new Error(message ?? constMessage(value))
+    throw new Error(message ?? constantMessage(value))
   })
 }